Fix: 카카오 도서 API 버그 수정 및 ISBN int→long 타입 변환
- KakaoBookController: NPE 방지(getErrorStream null 체크), 리소스 누수 수정(try-with-resources)
- AdminDAO: INSERT 시 컬럼명 명시하여 category 컬럼 불일치 오류 수정
- ISBN 타입 int→long 전체 변환 (VO, DAO, Service, Controller, DB스키마 BIGINT)

+++ src/main/java/WebBookStore/admin/AdminController.java
@@ -36,7 +36,7 @@ public class AdminController {
 	}
 
 	@RequestMapping("/updateform")
-	public String updateform(@RequestParam("isbn") int isbn, Model model) {
+	public String updateform(@RequestParam("isbn") long isbn, Model model) {
 		model.addAttribute("admin", adminService.getBookById(isbn));
 		model.addAttribute("contentPage", "/WEB-INF/views/admin/updateform.jsp");
 		return "layout/layout";
@@ -53,7 +53,7 @@ public class AdminController {
 	}
 
 	@RequestMapping("/delete")
-	public String delete(@RequestParam("isbn") int isbn) { // id 대신 isbn으로 명시
+	public String delete(@RequestParam("isbn") long isbn) {
 		adminService.deleteBook(isbn);
 		return "redirect:/book/list";
 	}
